Refactor of Utils for consolidation
